[android] Remove Google Play Location Services
GMS returns total crap instead of location. There is absolutely no control
on the source of location returned by FusedLocationClient. It can return
inaccurate (~700 meters) location from the network and mark it as "fused".
