--- android/flavors/gms-disabled/com/mapswithme/maps/location/LocationProviderFactory.java
@@ -1,18 +1,0 @@
-package com.mapswithme.maps.location;
-
-import android.content.Context;
-
-import androidx.annotation.NonNull;
-
-public class LocationProviderFactory
-{
-  public static boolean isGoogleLocationAvailable(@NonNull @SuppressWarnings("unused") Context context)
-  {
-    return false;
-  }
-
-  public static BaseLocationProvider getProvider(@NonNull Context context, @NonNull BaseLocationProvider.Listener listener)
-  {
-    return new AndroidNativeProvider(context, listener);
-  }
-}
