+++ android/res/values/donottranslate.xml
@@ -10,7 +10,6 @@
   <string name="pref_munits" translatable="false">MeasurementUnits</string>
   <string name="pref_storage" translatable="false">Storage</string>
   <string name="pref_show_zoom_buttons" translatable="false">ZoomButtonsEnabled</string>
-  <string name="pref_play_services" translatable="false">PlayServicesEnabled</string>
   <string name="pref_enable_logging" translatable="false">EnableLogging</string>
   <string name="pref_emulate_bad_external_storage" translatable="false">EmulateBadExternalStorage</string>
   <string name="pref_about" translatable="false">AboutOrganicMaps</string>
