+++ data/strings/strings.txt
@@ -5661,42 +5661,6 @@
     zh-Hans = 更新地图
     zh-Hant = 更新地圖
 
-  [pref_use_google_play]
-    comment = Preference text
-    tags = android
-    en = Use Google Play Services to determine your current location
-    ar = استخدم خدمات جوجل بلاي لتتعرف على موقعك الحالي
-    be = Ужываць службы Google Play для вызначэння месцазнаходжання
-    bg = Използвайте услугите на Google Play, за да определите текущото си местоположение
-    cs = Používat Služby Google Play pro získání aktuální polohy
-    da = Anvend Google Play Services til at få din aktuelle placering
-    de = Nutzen Sie die Google Play-Dienste, um Ihren aktuellen Standort zu erhalten
-    el = Χρησιμοποιήστε υπηρεσίες Google Play για να προσδιορίσετε την τρέχουσα τοποθεσία σας
-    es = Usar los servicios de Google Play para obtener tu ubicación actual
-    fa = استفاده از Google Play Services برای تعیین موقعیت کنونی شما
-    fi = Käytä Google Play -palveluita sijaintisi selvittämiseksi
-    fr = Utilisez l'application Google Play Services pour obtenir votre emplacement actuel
-    hu = Használja a Google Play szolgáltatást, hogy szert tegyen aktuális helyszínéről
-    id = Gunakan Layanan Google Play untuk mendapatkan lokasi Anda saat ini
-    it = Usa i servizi Google Play per ottenere la tua posizione attuale
-    ja = 現在のロケーションを取得するためにGoogle Playのサービスを使う
-    ko = Google Play 서비스를 사용하여 현재 위치 정보 얻기
-    nb = Bruk Google Play Services for å hente din nåværende posisjon
-    nl = Gebruik de diensten van Google Play om uw huidige locatie te bepalen
-    pl = Używa usług Google Play do ustalenia aktualnego położenia
-    pt = Use os Serviços Google Play para determinar a sua localização atual
-    pt-BR = Usar Serviços do Google Play para determinar a sua localização atual
-    ro = Folosește serviciile Google Play, pentru a obține poziția actuală
-    ru = Использовать Google Play Services для определения позиции
-    sk = Pomocou Google Play služieb získajte svoju aktuálnu polohu
-    sv = Använd Google Play Services för att bestämma din aktuella position
-    th = ใช้บริการ Google Play เพื่อรับตำแหน่งปัจจุบันของคุณ
-    tr = Mevcut konumunuzu belirlemek için Google Play Hizmetlerini kullanın
-    uk = Щоб визначити ваше поточне місцезнаходження, використовуйте сервіси Google Play
-    vi = Sử dụng Dịch vụ Google Play để có được vị trí hiện tại của bạn
-    zh-Hans = 使用Google Play服务以获取您的当前位置。
-    zh-Hant = 使用 Google Play 服務來獲得您的目前位置
-
   [routing_download_maps_along]
     comment = Text for routing error dialog
     tags = android,ios
